# [refact] use cute::tensor to define smem_d layout