@LyricZhao 2025-07-30
Sorry, I test your PR on H800 and the performance is lower than before (e.g. m= 4096, n= 7168, k=16384, 1D2D, layout=NT, BF16, acc=0, 1500 TFLOPS v.s. 1473 TFLOPS).

Personally, I am not a big fan of CuTe, as we will miss lots of detailed optimizations. For the swizzling part, without CuTe:

- we can control all the dtype as 32 bit (I am not sure CuTe will use 64 bit, which leads to more regs)
- we can do case expression bypass, see `constexpr bool kHasShortcut = (kSwizzleDMode / kNumBankGroupBytes) == 8;`
- no used to use `threadIdx.x` for several times
- we can control the instruction order, which may lead to different performance
- so on...

In my opinion, CUTLASS/CuTe routine maybe could provide 90-99% performance, but not 100%. Currently I don't think it is tedious to write (as optimizing every single detail is very interesting). For easier maintainence, we may abstract our impl as a struct later.